fix backup when file not exits

+++ src/main/java/team/ebi/epicbanitem/configuration/ObservableConfigFile.java
@@ -7,6 +7,7 @@ import ninja.leaping.configurate.loader.ConfigurationLoader;
 import team.ebi.epicbanitem.EpicBanItem;
 import team.ebi.epicbanitem.util.file.ObservableFile;
 
+import javax.annotation.Nullable;
 import java.io.Closeable;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -53,10 +54,14 @@ public class ObservableConfigFile implements ObservableFile, Closeable {
                 this.updateConsumer.accept(node);
             }
         } catch (Exception e) {
-            EpicBanItem.getLogger().warn("Failed to load config file {}, a backup is created at {}", path, backup);
+            if (backup != null) {
+                EpicBanItem.getLogger().warn("Failed to load config file {}, a backup is created at {}", path, backup);
+            }
             throw e;
         }
-        Files.delete(backup);
+        if (backup != null) {
+            Files.delete(backup);
+        }
     }
 
     public void save() throws IOException {
@@ -70,7 +75,11 @@ public class ObservableConfigFile implements ObservableFile, Closeable {
         }
     }
 
+    @Nullable
     public Path backup() throws IOException {
+        if (!Files.exists(this.path)) {
+            return null;
+        }
         Path backupDir = configDir.resolve("backup");
         Path aPath = backupDir.resolve(configDir.relativize(this.path));
         Path dir = aPath.getParent();
